Fixing mainly the `collector` table. Making the email and facebook_id unique fields

+++ lib/model/map/CollectionCategoryTableMap.php
@@ -38,7 +38,7 @@ class CollectionCategoryTableMap extends TableMap
     $this->setPackage('lib.model');
     $this->setUseIdGenerator(true);
     // columns
-    $this->addPrimaryKey('ID', 'Id', 'SMALLINT', true, null, null);
+    $this->addPrimaryKey('ID', 'Id', 'INTEGER', true, null, null);
     $this->addColumn('PARENT_ID', 'ParentId', 'INTEGER', false, null, 0);
     $this->addColumn('NAME', 'Name', 'VARCHAR', true, 64, null);
     $this->getColumn('NAME', false)->setPrimaryString(true);
